+++ tests/specs/prompt_parser_spec.lua
@@ -2,7 +2,7 @@
 --
 -- SPDX-License-Identifier: MIT
 
-local prompt_parser = require("telescope-live-grep-raw.prompt_parser")
+local prompt_parser = require("telescope-live-grep-args.prompt_parser")
 
 local tests = {
   {
